+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Fixed `MINIO_ACCESS_KEY_ID` environment variable for MinIO storage credentials.
 - Refactored XvcConfig for a simpler structure using structs for all configuration
 - Added documentation for configuration structs and their fields in `config/src/configuration.rs`.
+- Updated `workdir` handling in CLI to use absolute paths and default to the current directory if not specified.
+- Fixed a bug in configuration file generation where `[check-ignore]` section was incorrectly named `[check_ignore]`.
 
 
 ## v0.6.17 (2025-04-22)